+++ src/briefy/common/db/mixins/timestamp.py
@@ -1,6 +1,8 @@
 """Timestamp mixin."""
+from briefy.common.config import IMPORT_KNACK
 from briefy.common.db import datetime_utcnow
 from briefy.common.db.types import AwareDateTime
+from sqlalchemy.ext.declarative import declared_attr
 
 import sqlalchemy as sa
 
@@ -20,13 +22,16 @@ class Timestamp:
     Returns a datetime object, always in UTC, representing when this object was created.
     """
 
-    updated_at = sa.Column(
-        AwareDateTime(),
-        default=datetime_utcnow,
-        # TODO: return this after migration
-        # onupdate=datetime_utcnow
-    )
-    """Last update date of this object.
+    @declared_attr
+    def updated_at(cls):
+        """Last update date of this object.
 
-    Returns a datetime object, always in UTC, representing when this object was last updated.
-    """
+        Returns a datetime object, always in UTC, representing when this object was last updated.
+        """
+        kwargs = dict(default=datetime_utcnow)
+        if not IMPORT_KNACK:
+            kwargs.update(onupdate=datetime_utcnow)
+        return sa.Column(
+            AwareDateTime(),
+            **kwargs,
+        )