+++ HISTORY.rst
@@ -34,6 +34,8 @@ History
     * Add new mixin to hold contact information: company_name, email, mobile and additional_phone (rudaporto).
     * Change event name validator to accept names with _ (underscore) (rudaporto).
     * Fix bug in the workflow._perform_transition, it should use the name of the transition not the title (rudaporto).
+    * New config.IMPORT_KNACK to be used by the import script: disable Timestamp.updated_at onupdate parameter (rudaporto).
+    * Makefile improvements (rudaporto).
  
 
 1.1.2 (2016-10-04)